requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
alembic==1.12.1
pydantic==2.5.0
aiofiles==23.2.1
python-dotenv==1.0.0
asyncpg==0.29.0
websockets==12.0
redis==5.0.1